requirements.txt
numpy>=1.26
scipy>=1.12
pandas>=2.2
matplotlib>=3.8
PyYAML>=6.0
notebook>=7.0
ipykernel>=6.29
pytest>=8.0
nbformat>=5.10
